In evaluate, only set serialization_options if return_by_value=False (#196)
* In evaluate, only set serialization_options if return_by_value=False

Fixes #195: now calling evaluate with return_by_value=True and an
expression that evaluates to a JSON object will return the
corresponding Python object.

A consequence of this change is that calling evaluate with
return_by_value=True and an expression that evalutates to a non-JSON
result now result in a ProtocolException. Callers may need to update
their code to pass return_by_value=False.

* Update CHANGELOG.md for fix for #195.